ANCIEN BACK/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.7.0
cachetools==5.3.1
cffi==1.17.1
chardet==5.1.0
click==8.1.6
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.2
distlib==0.3.7
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
filelock==3.12.2
flake8==6.0.0
flake8-black==0.3.6
flake8-bugbear==23.7.10
flake8-docstrings==1.7.0
flake8-typing-imports==1.14.0
greenlet==3.0.1
iniconfig==2.0.0
isort==5.12.0
lxml==5.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.2
openpyxl==3.1.5
packaging==23.1
pandas==2.1.3
pathspec==0.11.2
pep8-naming==0.13.3
platformdirs==3.9.1
pluggy==1.2.0
protobuf==4.21.12
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
PyJWT==2.8.0
pyproject-api==1.5.3
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
snowballstemmer==2.2.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
tomli==2.0.1
tox==4.6.4
typing_extensions==4.6.3
tzdata==2023.3
virtualenv==20.24.1
